# Use AVX2 in `minmax_element` vectorization