@StephanTLavavej 2024-06-18
AVX2 fast, AVX2 furious! :car: :blue_car: :racing_car: